added in copy mode for moving of folders in the asset browser currently it is implementd as a check box to enable copy mode , by default when copy mode is unchecked it is move mode

+++ Eclipse/src/Editor/Utilities/DragAndDrop.cpp
@@ -56,8 +56,9 @@ namespace Eclipse
 	}
 	
 	void DragAndDrop::AssetBrowerFilesAndFoldersTarget(const char* type, const char* paths, 
-		std::string AssetPath, std::filesystem::directory_entry dirEntry, bool& refreshBrowser, std::map<std::filesystem::path, std::vector<std::filesystem::path>> pathMap)
+		std::string AssetPath, std::filesystem::directory_entry dirEntry, bool& refreshBrowser, std::map<std::filesystem::path, std::vector<std::filesystem::path>> pathMap, bool& copyFile)
 	{
+		
 		static std::string folderName;
 		
 		if (ImGui::BeginDragDropTarget())
@@ -181,14 +182,20 @@ namespace Eclipse
 										break;
 									}
 								}
-								for (auto const& it : deletefiles)
+								if (!copyFile)
 								{
-									std::filesystem::remove_all(std::filesystem::path(it.first.c_str()));
+									for (auto const& it : deletefiles)
+									{
+										std::filesystem::remove_all(std::filesystem::path(it.first.c_str()));
+									}
 								}
-								if (baseFile)
+								if(baseFile)
 								{
 									baseFile = false;
-									std::filesystem::remove(std::filesystem::path(parentPath));
+									if(!copyFile)
+									{
+										std::filesystem::remove(std::filesystem::path(parentPath));
+									}
 								}
 							}
 						}
@@ -205,7 +212,10 @@ namespace Eclipse
 							}
 							std::filesystem::copy(std::filesystem::path(itemPaths / paths), dirEntry.path().string() + "//" + folderName, copyOptions);
 						}
-						std::filesystem::remove_all(std::filesystem::path(itemPaths / paths));
+						if(!copyFile)
+						{
+							std::filesystem::remove_all(std::filesystem::path(itemPaths / paths));
+						}
 						//resetting for the next moving of files
 						deletefiles.clear();
 						files.clear();
@@ -216,7 +226,10 @@ namespace Eclipse
 					else
 					{
 						std::filesystem::copy(std::filesystem::path(itemPaths / paths), dirEntry.path());
-						std::filesystem::remove(std::filesystem::path(itemPaths / paths));
+						if(!copyFile)
+						{
+							std::filesystem::remove(std::filesystem::path(itemPaths / paths));
+						}
 					}
 					refreshBrowser = true;
 				}
